+++ src/io/XDMFStream.h
@@ -41,9 +41,9 @@ public:
     std::string path() const;
 
 
-    virtual std::tuple<bool, std::string> open(std::string const &url, size_t flag = 0UL);
+    virtual std::tuple<bool, std::string> open(std::string const &url, int flag = 0UL);
 
-    virtual std::string read(std::string const &url, data_model::DataSet *ds, size_t flag = 0UL) { return ""; };
+    virtual std::string read(std::string const &url, data_model::DataSet *ds, int flag = 0UL) { return ""; };
 
     void close();
 
@@ -60,7 +60,7 @@ public:
 
     void time(Real time);
 
-    std::string write(std::string const &s, data_model::DataSet const &ds, size_t flag = 0UL);
+    std::string write(std::string const &s, data_model::DataSet const &ds, int flag = 0UL);
 
 //    void write(std::string const &s, base::AttributeObject const &ds);
